requirements.txt
Django>=5.0,<6
psycopg[binary]>=3.1
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6
django-environ>=0.11
stripe>=8.0
